optimize drawing / fix types

+++ Source/Material/Icons.h
@@ -5,7 +5,7 @@
 namespace Material {
     class Icon {
     public:
-        Icon(const char *data, const size_t size)
+        Icon(const char *data, const int size)
                 : svg(XmlDocument::parse(String(CharPointer_UTF8(data), size))) {
         }
